Create ShopperInsights module and move files there.
