+++ ShopperInsights/src/main/java/com/braintreepayments/api/shopperinsights/BuyerPhone.kt
@@ -1,4 +1,4 @@
-package com.braintreepayments.api
+package com.braintreepayments.api.shopperinsights
 
 /**
  * Representation of a user phone number.
