+++ ShopperInsights/src/main/java/com/braintreepayments/api/shopperinsights/ShopperInsightsRequest.kt
@@ -1,4 +1,4 @@
-package com.braintreepayments.api
+package com.braintreepayments.api.shopperinsights
 
 /**
  * Data class representing a request for shopper insights.
