+++ ShopperInsights/src/main/java/com/braintreepayments/api/shopperinsights/ShopperInsightsResult.kt
@@ -1,4 +1,4 @@
-package com.braintreepayments.api
+package com.braintreepayments.api.shopperinsights
 
 /**
  * The result object returned when insights about a shopper is requested.
