+++ settings.gradle
@@ -16,3 +16,4 @@ include ':TestUtils'
 include ':Demo'
 include ':SEPADirectDebit'
 include ':PayPalDataCollector'
+include ':ShopperInsights'
